fix: DH-21298: Factor out test utility classes
