+++ src/test/java/io/deephaven/csv/testutil/MakeCustomColumn.java
@@ -1,0 +1,5 @@
+package io.deephaven.csv.testutil;
+
+public interface MakeCustomColumn {
+    Column<?> apply(String name, Object column, int size);
+}
