+++ src/test/java/io/deephaven/csv/testutil/Sentinels.java
@@ -1,0 +1,14 @@
+package io.deephaven.csv.testutil;
+
+public class Sentinels {
+    public static final byte NULL_BOOLEAN_AS_BYTE = Byte.MIN_VALUE;
+    public static final byte NULL_BYTE = Byte.MIN_VALUE;
+    public static final short NULL_SHORT = Short.MIN_VALUE;
+    public static final int NULL_INT = Integer.MIN_VALUE;
+    public static final long NULL_LONG = Long.MIN_VALUE;
+    public static final float NULL_FLOAT = -Float.MAX_VALUE;
+    public static final double NULL_DOUBLE = -Double.MAX_VALUE;
+    public static final char NULL_CHAR = Character.MAX_VALUE;
+    public static final long NULL_DATETIME_AS_LONG = Long.MIN_VALUE;
+    public static final long NULL_TIMESTAMP_AS_LONG = Long.MIN_VALUE;
+}
